+++ CHANGES.md
@@ -3,11 +3,14 @@ Changes by Version
 Release Notes.
 
 0.8.0
+------------------
 
 ### Features
 
 * Bump up the API to support the index mode of Measure.
 * Bump up the API to support the new property.
+* Bump up the API to adopt the status field which is changed to the string type due to the compatibility issue.
+* Bump up the API to support getting the API version.
 
 0.7.0
 ------------------